requirements-dev.txt
-r requirements.txt
black
mypy
pandas-stubs
pylint
pytest >= 4.1.1
pytest-cov >= 2.6.1